Use ReceiverMock instead create new in dispatcher
