+++ data-plane/receiver/src/test/java/dev/knative/eventing/kafka/broker/receiver/MockReactiveKafkaProducer.java
@@ -16,7 +16,6 @@
 package dev.knative.eventing.kafka.broker.receiver;
 
 import dev.knative.eventing.kafka.broker.core.ReactiveKafkaProducer;
-import io.cloudevents.CloudEvent;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
 import java.util.Properties;
@@ -25,32 +24,20 @@ import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 
-public class MockReactiveKafkaProducer<K, V> implements ReactiveKafkaProducer<String, CloudEvent> {
+public class MockReactiveKafkaProducer<K, V> implements ReactiveKafkaProducer<K, V> {
 
-    private final Producer<String, CloudEvent> producer;
+    private final Producer<K, V> producer;
+
+    public MockReactiveKafkaProducer(Producer<K, V> producer) {
+        this.producer = producer;
+    }
 
     public MockReactiveKafkaProducer(Properties config) {
         this.producer = new KafkaProducer<>(config);
     }
 
-    public MockReactiveKafkaProducer(Producer<String, CloudEvent> producer) {
-        this.producer = producer;
-    }
-
     @Override
-    public Future<Void> close() {
-        producer.close();
-        return Future.succeededFuture();
-    }
-
-    @Override
-    public Future<Void> flush() {
-        producer.flush();
-        return Future.succeededFuture();
-    }
-
-    @Override
-    public Future<RecordMetadata> send(ProducerRecord<String, CloudEvent> record) {
+    public Future<RecordMetadata> send(ProducerRecord<K, V> record) {
         final Promise<RecordMetadata> p = Promise.promise();
         producer.send(record, (recordMetadata, exception) -> {
             if (exception != null) {
@@ -63,7 +50,19 @@ public class MockReactiveKafkaProducer<K, V> implements ReactiveKafkaProducer<St
     }
 
     @Override
-    public org.apache.kafka.clients.producer.Producer<String, CloudEvent> unwrap() {
+    public Future<Void> close() {
+        producer.close();
+        return Future.succeededFuture();
+    }
+
+    @Override
+    public Producer<K, V> unwrap() {
         return producer;
+    }
+
+    @Override
+    public Future<Void> flush() {
+        producer.flush();
+        return Future.succeededFuture();
     }
 }
