+++ data-plane/receiver/src/test/java/dev/knative/eventing/kafka/broker/receiver/MockReactiveProducerFactory.java
@@ -17,15 +17,14 @@ package dev.knative.eventing.kafka.broker.receiver;
 
 import dev.knative.eventing.kafka.broker.core.ReactiveKafkaProducer;
 import dev.knative.eventing.kafka.broker.core.ReactiveProducerFactory;
-import io.cloudevents.CloudEvent;
 import io.vertx.core.Vertx;
 import java.util.Properties;
 import org.apache.kafka.clients.producer.MockProducer;
 
-public class MockReactiveProducerFactory implements ReactiveProducerFactory<String, CloudEvent> {
+public class MockReactiveProducerFactory<K, V> implements ReactiveProducerFactory<K, V> {
 
     @Override
-    public ReactiveKafkaProducer<String, CloudEvent> create(Vertx v, Properties properties) {
-        return new MockReactiveKafkaProducer<String, CloudEvent>(new MockProducer<>(true, null, null));
+    public ReactiveKafkaProducer<K, V> create(Vertx v, Properties configs) {
+        return new MockReactiveKafkaProducer<>(new MockProducer<>(true, null, null));
     }
 }
